requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-texttospeech==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
packaging==20.9
proto-plus==1.18.1
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyTelegramBotAPI==3.7.7
python-slugify==5.0.2
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
